Log the fact of successful plugin load
Unfortunately Burp Suite API doesn't provide any facilities to check plugin states in runtime. The only way to automate these things is to monitor output of the plugin, but Retire.js doesn't report anything useful after being loaded. It wouldn't be a problem if the plugin loaded immediately, but it preloads vulnerabilities from outside and it could take a while. This commit just prints a message to the stdout after plugin load.